final-back/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0
django-allauth==0.40.0
django-cors-headers==3.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
